feat(v3.15.0): fix hx-on chet + sieu bao mat + sieu toc (khong doi UI)
- hx-on::after-request dang ky sai case -> handler khong bao gio chay:
  sua 5 diem + form chat dung listener JS (giu chu khi gui loi).
- Bao mat: WS cap 64KB->128KB (unlimited 20k ky tu); headers moi
  X-DNS-Prefetch-Control/X-Permitted-Cross-Domain-Policies;
  Referrer ve same-origin; throttle brute-force AI login 20/15phut;
  thu hoi session cu khi login (chong fixation).
- Toc do: fix heatmap date-bigint (query loi moi luot xem ho so);
  preconnect avatar Google.
- Gates: check/clippy -D warnings/test 389/doc/fmt OK.

+++ src/repositories/activity.rs
@@ -47,6 +47,11 @@ impl ActivityRepo {
     /// Đồng bộ: `(NOW() AT TIME ZONE 'Asia/Ho_Chi_Minh')::date` — đúng
     /// quy ước dự án, không phụ thuộc TZ cài đặt của DB.
     pub async fn heatmap(pool: &PgPool, user_id: Uuid) -> AppResult<Vec<HeatmapDay>> {
+        // v3.15.0 FIX (DB error mỗi lượt xem hồ sơ): `::date - $2` với $2
+        // kiểu i64 (int8) → Postgres KHÔNG có operator `date - bigint`
+        // (chỉ có `date - integer`) → query fail, heatmap luôn rỗng +
+        // spam log `operator does not exist: date - bigint`. Bind i32
+        // (int4) để dùng đúng operator có sẵn.
         let rows = sqlx::query_as::<_, HeatmapDay>(
             "SELECT day, activity_count FROM user_activity_days
              WHERE user_id = $1
@@ -54,7 +59,7 @@ impl ActivityRepo {
              ORDER BY day ASC",
         )
         .bind(user_id)
-        .bind(HEATMAP_DAYS)
+        .bind(HEATMAP_DAYS as i32)
         .fetch_all(pool)
         .await?;
         Ok(rows)
